Moved processing code to LessRuleset, ParameterMixin, ParameterRuleset and UnprocessedStatement. And using polymorphism rather than downcasting StylesheetStatement and RulesetStatement implementations.
StylesheetStatements and RulesetStatements implement a process()
function to process their code and insert the result into the provided
argument. This leaves a LessStylesheet unchanged after processing and
creates a separate css-only stylesheet to write to output.

They also implement a write(CssWriter*) function which is used to
instruct the CssWriter how to output the code.