Fixed importing so Token.source is a pointer to a string in the sources list.
When a new less stylesheet is imported, a new string is generated that
contains the file path, relative to the initial source file. That
string is stored in the sources list that the LessParser passes
along. When parsing is done the sources list contains the paths to all
the source files.

The same string that goes into the list also is passed to the
LessTokenizer which then assigns it to all the Tokens it
generates. The Token.source pointer can be matched against an entry in
the sources list.

All this is necessary for the SourceMapWriter to accurately output the
index of a source in the sources list.