Updating the way mixins and variables are found.
Previously rulesets called by mixins were found by looking in the same
ruleset the mixin is in, and then that ruleset's parent, and so
on. Variables were stored using a scope stack.

The right way to find the rulesets that match a mixin or the value of
a variable is to look at the mixin's/variable's ruleset, that
ruleset's parent, etc. If nothing is found that way, the same
procedure happens for the last ruleset on the call stack.

Example:

.namespace1 {
  .local() {
     background: green;
  }
  @local-var: blue;

  .caller()  {
     .local();
     color: @local-var;
  }
}

.test {
  .local () {
     background: red;
  }
  @local-var: purple;

  .namespace1 .caller();
}

yields:

.test {
  background: green;
  color: blue;
}