Rewrote ProcessingContext to find mixins and variables the right way.
ValueScope has been changed to an interface, that is implemented by
ProcessingContext. Instead of a scope, a stack is used. Each level of
the stack contains a reference to a Function and a VariableMap.

Example:

.mixin(@x) {
   .nested {
     x:@x;
   }
}
.test {
  .mixin(1);
}

At the declaration 'x:@x' the stack will be '.test, ()' ->
'.mixin, (@x=5)' -> '.nested, ()'

Added VariableMap to extend std::map<std::string, const TokenList>.